Move classes into the Hydra::Works::Characterization namespace
