--- lib/hydra/works/models/characterization/video_schema.rb
@@ -1,7 +1,0 @@
-class VideoSchema < ActiveTriples::Schema
-  property :height, predicate: RDF::Vocab::EBUCore.height
-  property :width, predicate: RDF::Vocab::EBUCore.width
-  property :frame_rate, predicate: RDF::Vocab::NFO.frameRate
-  property :duration, predicate: RDF::Vocab::NFO.duration
-  property :sample_rate, predicate: RDF::Vocab::EBUCore.sampleRate
-end
