+++ lib/hydra/works/models/concerns/file_set/characterization/audio.rb
@@ -2,8 +2,6 @@ module Hydra::Works::Characterization
   module Audio
     extend ActiveSupport::Concern
 
-    autoload :AudioSchema, 'hydra/works/models/characterization/audio_schema.rb'
-
     included do
       # Apply the audio schema. This will add properties defined in the schema.
       apply_schema AudioSchema, AlreadyThereStrategy
