+++ lib/hydra/works/models/concerns/file_set/characterization/video.rb
@@ -2,8 +2,6 @@ module Hydra::Works::Characterization
   module Video
     extend ActiveSupport::Concern
 
-    autoload :VideoSchema, 'hydra/works/models/characterization/video_schema.rb'
-
     included do
       # Apply the module schema. This will add properties defined in the schema.
       apply_schema VideoSchema, AlreadyThereStrategy
